Struct cpp_to_rust::cpp_data::CppData

pub struct CppData {
    pub types: Vec<CppTypeData>,
    pub methods: Vec<CppMethod>,
    pub template_instantiations: Vec<CppTemplateInstantiations>,
}

C++ parser output

Fields
types: Vec<CppTypeData>

List of found type declarations

methods: Vec<CppMethod>

List of found methods

template_instantiations: Vec<CppTemplateInstantiations>

List of found template instantiations. Key is name of the template class, value is list of instantiations.

Methods

impl CppData

fn ensure_explicit_destructors(&mut self, dependencies: &[&CppData]) -> Result<()>

Adds destructors for every class that does not have explicitly defined destructor, allowing to create wrappings for destructors implicitly available in C++.

fn add_inherited_methods(&mut self, dependencies: &[&CppData]) -> Result<()>

Adds methods of derived classes inherited from base classes. A method will not be added if there is a method with the same name in the derived class. Constructors, destructors and assignment operators are also not added. This reflects C++'s method inheritance rules.

fn generate_methods_with_omitted_args(&mut self)

Generates duplicate methods with fewer arguments for C++ methods with default argument values.

fn all_include_files(&self) -> HashSet<String>

fn is_template_class(&self, name: &str) -> bool

Checks if specified class is a template class.

fn has_virtual_destructor(&self, class_name: &str, dependencies: &[&CppData]) -> bool

Checks if specified class has virtual destructor (own or inherited).

fn has_public_destructor(&self, class_type: &CppTypeClassBase) -> bool

Checks if specified class has public destructor.

fn get_all_methods(&self, class_name: &str) -> Result<Vec<&CppMethod>>

fn has_pure_virtual_methods(&self, class_name: &str) -> bool

fn add_field_accessors(&mut self) -> Result<()>

fn post_process(&mut self, dependencies: &[&CppData]) -> Result<()>

Performs data conversion to make it more suitable for further wrapper generation.