Struct cpp_to_rust::cpp_method::CppFunctionArgument

pub struct CppFunctionArgument {
    pub name: String,
    pub argument_type: CppType,
    pub has_default_value: bool,
}

Information about an argument of a C++ method

Fields
name: String

Identifier. If the argument doesn't have a name (which is allowed in C++), this field contains generated name "argX" (X is position of the argument).

argument_type: CppType

Argument type

has_default_value: bool

Flag indicating that the argument has default value and therefore can be omitted when calling the method