Struct cpp_to_rust::cpp_method::CppMethod

pub struct CppMethod {
    pub name: String,
    pub class_membership: Option<CppMethodClassMembership>,
    pub operator: Option<CppOperator>,
    pub return_type: CppType,
    pub arguments: Vec<CppFunctionArgument>,
    pub arguments_before_omitting: Option<Vec<CppFunctionArgument>>,
    pub allows_variadic_arguments: bool,
    pub include_file: String,
    pub origin_location: Option<CppOriginLocation>,
    pub template_arguments: Option<TemplateArgumentsDeclaration>,
    pub template_arguments_values: Option<Vec<CppType>>,
    pub declaration_code: Option<String>,
    pub inheritance_chain: Vec<CppBaseSpecifier>,
    pub doc: Option<CppMethodDoc>,
}

Information about a C++ method

Fields
name: String

Identifier. For class methods, this field includes only the method's own name. For free functions, this field also includes namespaces (if any).

class_membership: Option<CppMethodClassMembership>

Additional information about a class member function or None for free functions

operator: Option<CppOperator>

If the method is a C++ operator, indicates its kind

return_type: CppType

Return type of the method. Return type is reported as void for constructors and destructors.

arguments: Vec<CppFunctionArgument>

List of the method's arguments

arguments_before_omitting: Option<Vec<CppFunctionArgument>>

If Some, the method is derived from another method by omitting arguments, and this field contains all arguments of the original method.

allows_variadic_arguments: bool

Whether the argument list is terminated with "..."

include_file: String

File name of the include file where the method is defined (without full path)

origin_location: Option<CppOriginLocation>

Exact location of declaration of the method. Can be None if the method is generated automatically and doesn't have corresponding C++ declaration.

template_arguments: Option<TemplateArgumentsDeclaration>

Names of the method's template arguments. None if this is not a template method. If the method belongs to a template class, the class's template arguments are not included here.

template_arguments_values: Option<Vec<CppType>>

For an instantiated template method, this field contains the types used for instantiation. For example, T QObject::findChild<T>() would have no template_arguments_values because it's not instantiated, and QWidget* QObject::findChild<QWidget*>() would have QWidget* type in template_arguments_values.

declaration_code: Option<String>

C++ code of the method's declaration. None if the method was not explicitly declared.

inheritance_chain: Vec<CppBaseSpecifier>

List of base classes this method was inferited from. The first item is the most base class.

doc: Option<CppMethodDoc>

C++ documentation data for this method

Methods

impl CppMethod

fn argument_types_equal(&self, other: &CppMethod) -> bool

Checks if two methods have exactly the same set of input argument types

fn needs_allocation_place_variants(&self) -> bool

Checks if this method would need to have 2 wrappers with 2 different return value allocation places

fn c_signature(&self, allocation_place: ReturnValueAllocationPlace) -> Result<CppFfiFunctionSignature>

Creates FFI method signature for this method: - converts all types to FFI types; - adds "this" argument explicitly if present; - adds "output" argument for return value if allocation_place is Stack.

fn to_ffi_signatures(&self) -> Result<Vec<CppMethodWithFfiSignature>>

Generates either one or two FFI signatures for this method, depending on its return type.

fn full_name(&self) -> String

fn doc_id(&self) -> String

fn short_text(&self) -> String

Returns short text representing values in this method (only for debug output purposes).

fn inheritance_chain_text(&self) -> String

fn class_name(&self) -> Option<&String>

fn is_constructor(&self) -> bool

fn is_destructor(&self) -> bool

fn class_info_if_constructor(&self) -> Option<&CppMethodClassMembership>

fn is_operator(&self) -> bool

fn all_involved_types(&self) -> Vec<CppType>