Struct cpp_to_rust::cpp_method::CppMethodClassMembership

pub struct CppMethodClassMembership {
    pub class_type: CppTypeClassBase,
    pub kind: CppMethodKind,
    pub is_virtual: bool,
    pub is_pure_virtual: bool,
    pub is_const: bool,
    pub is_static: bool,
    pub visibility: CppVisibility,
    pub is_signal: bool,
    pub field_accessor: Option<CppFieldAccessor>,
}

Information about a C++ class member method

Fields
class_type: CppTypeClassBase

Type of the class where this method belong. This is used to construct type of "this" pointer and return type of constructors.

kind: CppMethodKind

Whether this method is a constructor, a destructor or an operator

is_virtual: bool

True if this is a virtual method

is_pure_virtual: bool

True if this is a pure virtual method (requires is_virtual = true)

is_const: bool

True if this is a const method, i.e. "this" pointer receives by this method has const type

is_static: bool

True if this is a static method, i.e. it doesn't receive "this" pointer at all.

visibility: CppVisibility

Method visibility

is_signal: bool

True if the method is a Qt signal

field_accessor: Option<CppFieldAccessor>

If this method is a generated field accessor, this field contains information about it. Field accessors do not have real C++ methods corresponding to them.