Struct cpp_to_rust::cpp_type::CppType

pub struct CppType {
    pub base: CppTypeBase,
    pub indirection: CppTypeIndirection,
    pub is_const: bool,
    pub is_const2: bool,
}

Information about a C++ type

Fields
base: CppTypeBase

Information about base type

indirection: CppTypeIndirection

Indirection applied to base type

is_const: bool

If the type has const qualifier. Defaults to false when not applicable.

is_const2: bool

If 2nd indirection part of the type is const, e.g. true for "int* const*".

Methods

impl CppType

fn void() -> Self

fn is_void(&self) -> bool

fn to_cpp_code(&self, function_pointer_inner_text: Option<&String>) -> Result<String>

fn to_cpp_pseudo_code(&self) -> String

fn to_cpp_ffi_type(&self, role: CppTypeRole) -> Result<CppFfiType>

Converts this C++ type to its adaptation for FFI interface, removing all features not supported by C ABI (e.g. references and passing objects by value)

fn caption(&self, strategy: TypeCaptionStrategy) -> Result<String>

Generates alphanumeric representation of self used to generate FFI function names

fn needs_allocation_place_variants(&self) -> bool

Checks if a function with this return type would need to have 2 wrappers with 2 different return value allocation places

fn instantiate(&self, nested_level1: i32, template_arguments1: &[CppType]) -> Result<CppType>